Remove trailing white spaces from deb
I don't like trailing whitespaces in source files I work with.   I could
have done all of the files in this repository but only chose to do the
ones I care about to help limit the size of the PR.  I replaced the
trailing whitespace using the following commands.

    find deb -type f -exec sed -i.bak -E 's/[[:space:]]+$//' {} \;
    git clean -xfd

If you wish to further do this on all files in this repository then
simply execute:

    find * -type f -exec sed -i.bak -E 's/[[:space:]]+$//' {} \;
    git clean -xfd